modify the quentity to the decimal and then create the calculator

+++ Desktop Windwos form application/AvalableStock.cs
@@ -12,7 +12,7 @@ namespace Desktop_Windwos_form_application
 
         public decimal Price { get; set; }
         public int ProductId { get; set; }
-        public int StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
         public string ProductName { get; set; }
     }
 }
